requirements.txt
openai
langchain
llama-parse
python-dotenv
pypdf
pdfplumber
pymupdf4llm
anthropic
pdf2image
pymupdf-layout